# swagger-spec.json
{"openapi":"3.0.0","paths":{"/users":{"get":{"operationId":"UsersController_findAllUser","parameters":[],"responses":{"200":{"description":""}},"tags":["User"]}},"/auth/register":{"post":{"operationId":"AuthController_createUser","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":""}},"tags":["Auth"]}},"/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthCredentialsDto"}}}},"responses":{"201":{"description":""}},"tags":["Auth"]}},"/posts":{"get":{"operationId":"PostsController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Posts"]},"post":{"operationId":"PostsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostDto"}}}},"responses":{"201":{"description":""}},"tags":["Posts"]}},"/posts/{id}":{"get":{"operationId":"PostsController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Posts"]},"patch":{"operationId":"PostsController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdatePostDto"}}}},"responses":{"200":{"description":""}},"tags":["Posts"]},"delete":{"operationId":"PostsController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Posts"]}},"/posts/slug/{slug}":{"get":{"operationId":"PostsController_getOneBySlug","parameters":[{"name":"slug","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Posts"]}},"/categories":{"get":{"operationId":"CategoriesController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Categories"]},"post":{"operationId":"CategoriesController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CategoriesDto"}}}},"responses":{"201":{"description":""}},"tags":["Categories"]}},"/categories/{id}":{"get":{"operationId":"CategoriesController_getOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Categories"]},"patch":{"operationId":"CategoriesController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCategoriesDto"}}}},"responses":{"200":{"description":""}},"tags":["Categories"]},"delete":{"operationId":"CategoriesController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Categories"]}},"/categories/list":{"get":{"operationId":"CategoriesController_getCategoriesList","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"array","items":{"type":"string"}}}}},"responses":{"200":{"description":""}},"tags":["Categories"]}},"/tags":{"get":{"operationId":"TagsController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Tags"]},"post":{"operationId":"TagsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTagDto"}}}},"responses":{"201":{"description":""}},"tags":["Tags"]}},"/tags/{id}":{"get":{"operationId":"TagsController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Tags"]},"patch":{"operationId":"TagsController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTagDto"}}}},"responses":{"200":{"description":""}},"tags":["Tags"]},"delete":{"operationId":"TagsController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Tags"]}},"/tags/list":{"post":{"operationId":"TagsController_getTagsList","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"array","items":{"type":"string"}}}}},"responses":{"201":{"description":""}},"tags":["Tags"]}},"/files/upload-image/{folderName}":{"post":{"operationId":"FilesController_uploadImage","parameters":[{"name":"folderName","required":true,"in":"path","schema":{"type":"string"}},{"name":"feature","required":true,"in":"query","schema":{"type":"boolean"}}],"responses":{"201":{"description":""}},"tags":["Files"]}}},"info":{"title":"TAKE ME AND SAVE ME API","description":"The TAKE ME AND SAVE ME API documentacion","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"CreateUserDto":{"type":"object","properties":{}},"AuthCredentialsDto":{"type":"object","properties":{}},"PostDto":{"type":"object","properties":{}},"UpdatePostDto":{"type":"object","properties":{}},"CategoriesDto":{"type":"object","properties":{}},"UpdateCategoriesDto":{"type":"object","properties":{}},"CreateTagDto":{"type":"object","properties":{}}}}}